split message types into separate files
